move pdf import settings into dialog
