+++ rnote-ui/src/settingspanel/mod.rs
@@ -36,12 +36,6 @@ mod imp {
         #[template_child]
         pub general_autosave_interval_secs_spinbutton: TemplateChild<SpinButton>,
         #[template_child]
-        pub general_pdf_import_width_adj: TemplateChild<Adjustment>,
-        #[template_child]
-        pub general_pdf_import_as_vector_toggle: TemplateChild<ToggleButton>,
-        #[template_child]
-        pub general_pdf_import_as_bitmap_toggle: TemplateChild<ToggleButton>,
-        #[template_child]
         pub general_format_border_color_choosebutton: TemplateChild<ColorButton>,
         #[template_child]
         pub format_predefined_formats_row: TemplateChild<adw::ComboRow>,
@@ -431,18 +425,6 @@ impl SettingsPanel {
         self.imp().settings_scroller.clone()
     }
 
-    pub fn general_pdf_import_width_adj(&self) -> Adjustment {
-        self.imp().general_pdf_import_width_adj.clone()
-    }
-
-    pub fn general_pdf_import_as_vector_toggle(&self) -> ToggleButton {
-        self.imp().general_pdf_import_as_vector_toggle.clone()
-    }
-
-    pub fn general_pdf_import_as_bitmap_toggle(&self) -> ToggleButton {
-        self.imp().general_pdf_import_as_bitmap_toggle.clone()
-    }
-
     pub fn general_format_border_color_choosebutton(&self) -> ColorButton {
         self.imp().general_format_border_color_choosebutton.clone()
     }
@@ -487,8 +469,6 @@ impl SettingsPanel {
     }
 
     pub fn load_general(&self, appwindow: &RnoteAppWindow) {
-        let pdf_import_as_vector = appwindow.canvas().engine().borrow().pdf_import_as_vector;
-        let pdf_import_width_perc = appwindow.canvas().engine().borrow().pdf_import_width_perc;
         let format_border_color = appwindow
             .canvas()
             .engine()
@@ -497,12 +477,6 @@ impl SettingsPanel {
             .format
             .border_color;
 
-        self.general_pdf_import_as_vector_toggle()
-            .set_active(pdf_import_as_vector);
-        self.general_pdf_import_as_bitmap_toggle()
-            .set_active(!pdf_import_as_vector);
-        self.general_pdf_import_width_adj()
-            .set_value(pdf_import_width_perc);
         self.general_format_border_color_choosebutton()
             .set_rgba(&gdk::RGBA::from_compose_color(format_border_color));
     }
@@ -621,28 +595,6 @@ impl SettingsPanel {
             .transform_from(|_, value| Some(f64::from(value.get::<u32>().unwrap()).to_value()))
             .flags(glib::BindingFlags::SYNC_CREATE | glib::BindingFlags::BIDIRECTIONAL)
             .build();
-
-        // Pdf import width
-        self.imp()
-            .general_pdf_import_width_adj
-            .get()
-            .connect_value_changed(clone!(@weak appwindow => move |general_pdf_import_width_adj| {
-                adw::prelude::ActionGroupExt::activate_action(&appwindow, "pdf-import-width-perc", Some(&general_pdf_import_width_adj.value().to_variant()));
-            }));
-
-        self.imp()
-            .general_pdf_import_as_vector_toggle
-            .get()
-            .connect_active_notify(clone!(@weak appwindow => move |general_pdf_import_as_vector_toggle| {
-                adw::prelude::ActionGroupExt::activate_action(&appwindow, "pdf-import-as-vector", Some(&general_pdf_import_as_vector_toggle.is_active().to_variant()));
-            }));
-
-        self.imp()
-            .general_pdf_import_as_bitmap_toggle
-            .get()
-            .connect_active_notify(clone!(@weak appwindow => move |general_pdf_import_as_bitmap_toggle| {
-                adw::prelude::ActionGroupExt::activate_action(&appwindow, "pdf-import-as-vector", Some(&(!general_pdf_import_as_bitmap_toggle.is_active()).to_variant()));
-            }));
 
         // revert format
         self.imp().format_revert_button.get().connect_clicked(
